Split off PanelManager from MainControlApplet.
Certain features (checkboxes for check trace, draw fields,
draw grid, show info) have been deactivated temporarily.
